+++ truelearn/datasets/_peek.py
@@ -21,7 +21,7 @@ from truelearn.models import (
     KnowledgeComponent,
     BaseKnowledgeComponent,
 )
-from ._base import RemoteFileMetaData, check_and_download_file
+from .base import RemoteFileMetaData, check_and_download_file
 
 
 PEEKData: TypeAlias = List[Tuple[int, List[Tuple[EventModel, bool]]]]
