+++ truelearn/learning/_interest_classifier.py
@@ -8,7 +8,7 @@ from truelearn.models import (
     LearnerModel,
     BaseKnowledgeComponent,
 )
-from ._base import (
+from .base import (
     InterestNoveltyKnowledgeBaseClassifier,
     gather_trueskill_team,
     team_sum_quality_from_kcs,
